feat(ui): network page (#2066)
* network page
* remove from settings
* add network status to network page
* refactor to improve code structure
* add license
* fixes based on review
* fixing part of the tests
* change copy
* Fix specs
