+++ ui/Screen/Settings.svelte
@@ -12,13 +12,7 @@
   import * as ethereum from "../src/ethereum";
   import * as ipc from "../src/ipc";
   import * as config from "../src/config";
-  import {
-    settings,
-    seedValidation,
-    addSeed,
-    removeSeed,
-    updateAppearance,
-  } from "../src/session";
+  import { settings, updateAppearance } from "../src/session";
   import * as sess from "../src/session";
   import {
     themeOptions,
@@ -31,12 +25,9 @@
 
   import {
     Button,
-    Icon,
     PeerId,
     SegmentedControl,
     SidebarLayout,
-    StyledCopyable,
-    TextInput,
   } from "ui/DesignSystem";
   import ModalShortcuts from "../Modal/Shortcuts.svelte";
 
@@ -53,18 +44,6 @@
     const environment = event.detail as ethereum.Environment;
     ethereum.selectedEnvironment.set(environment);
   };
-
-  let seedInputValue = "";
-
-  const submitSeed = async () => {
-    if (await addSeed(seedInputValue)) {
-      seedInputValue = "";
-    }
-  };
-
-  $: if (seedInputValue === "") {
-    seedValidation.reset();
-  }
 
   let version = "";
   (async () => {
@@ -130,50 +109,10 @@
     padding: 1rem 0.75rem;
   }
 
-  .section-item-single {
-    align-items: center;
-    padding: 1rem 0.75rem;
-  }
-
   .action {
     display: flex;
     justify-content: flex-end;
     margin-left: 1rem;
-  }
-
-  .seed-entry-form {
-    flex: 1;
-    display: flex;
-    flex-direction: column;
-    align-items: flex-start;
-  }
-
-  .seed-entry-field {
-    width: 100%;
-    display: flex;
-    align-items: flex-start;
-  }
-
-  .seeds {
-    display: flex;
-    flex-wrap: wrap;
-    justify-content: flex-end;
-    margin-top: 1.5rem;
-    width: 100%;
-  }
-
-  .seed {
-    width: 100%;
-    display: flex;
-    align-items: center;
-    justify-content: space-between;
-    border-top: 1px solid var(--color-foreground-level-2);
-    padding: 1.5rem 0.5rem;
-    cursor: default;
-  }
-
-  .seed:last-of-type {
-    border-bottom: 1px solid var(--color-foreground-level-2);
   }
 
   .title {
@@ -216,62 +155,6 @@
           <div class="action">
             <PeerId peerId={session.identity.peerId} />
           </div>
-        </div>
-      </section>
-
-      <section>
-        <header>
-          <h3>Network</h3>
-        </header>
-        <div class="section-item-single">
-          <p class="typo-text-bold">
-            Seeds help you find more projects and people on the network.
-          </p>
-          <p
-            style="color: var(--color-foreground-level-6); margin-bottom: 1.5rem;">
-            Enter seed addresses that you’d like to connect to here.
-            <a
-              style="color: var(--color-foreground-level-5);"
-              class="typo-link"
-              href="https://docs.radicle.xyz/docs/understanding-radicle/glossary#seed"
-              >Learn more</a>
-          </p>
-          <form
-            class="seed-entry-form"
-            on:submit|preventDefault
-            data-cy="seed-entry-form">
-            <div class="seed-entry-field">
-              <TextInput
-                dataCy="seed-input"
-                bind:value={seedInputValue}
-                placeholder="Enter a seed address here"
-                style="margin-right: 8px; min-width: 21.5rem; width: 100%;"
-                validation={$seedValidation} />
-              <Button
-                dataCy="add-seed"
-                style="display: flex;"
-                on:click={submitSeed}
-                disabled={!seedInputValue}
-                variant="outline">
-                Add
-              </Button>
-            </div>
-
-            <div class="seeds">
-              {#each $settings.coco.seeds as seed (seed)}
-                <div class="seed">
-                  <StyledCopyable value={seed} />
-                  <Button
-                    dataCy="remove-seed"
-                    style="padding:0.5rem;"
-                    on:click={() => removeSeed(seed)}
-                    variant="transparent">
-                    <Icon.Cross />
-                  </Button>
-                </div>
-              {/each}
-            </div>
-          </form>
         </div>
       </section>
